[PP][SUPP-224] moved journey common code to cor
