--- app/rootmodel/Url.scala
@@ -1,32 +1,0 @@
-/*
- * Copyright 2022 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package rootmodel
-
-import play.api.libs.functional.syntax._
-import play.api.libs.json.Format
-
-final case class BackUrl(value: String)
-
-object BackUrl {
-  implicit val format: Format[BackUrl] = implicitly[Format[String]].inmap(BackUrl(_), _.value)
-}
-
-final case class ReturnUrl(value: String)
-
-object ReturnUrl {
-  implicit val format: Format[ReturnUrl] = implicitly[Format[String]].inmap(ReturnUrl(_), _.value)
-}
