+++ cor-journey/src/main/scala/essttp/rootmodel/Email.scala
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-package journey
+package essttp.rootmodel
 
-import essttp.journey.model.JourneyId
+import play.api.libs.functional.syntax._
+import play.api.libs.json._
 
-import java.util.UUID
-import javax.inject.Singleton
+final case class Email(value: String)
 
-@Singleton
-class JourneyIdProvider {
-  def nextJourneyId(): JourneyId = JourneyId(UUID.randomUUID().toString)
+object Email {
+  implicit val format: Format[Email] = implicitly[Format[String]].inmap(Email(_), _.value)
 }
