+++ cor-journey/src/main/scala/essttp/rootmodel/SelectedPlan.scala
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-package journey
+package essttp.rootmodel
 
-import essttp.journey.model.JourneyId
+import play.api.libs.functional.syntax._
+import play.api.libs.json._
 
-import java.util.UUID
-import javax.inject.Singleton
+//TODO we dont know how the plan can be selected because we didn't agreed on the APIs with TTP yet
+final case class SelectedPlan(value: String)
 
-@Singleton
-class JourneyIdProvider {
-  def nextJourneyId(): JourneyId = JourneyId(UUID.randomUUID().toString)
+object SelectedPlan {
+  implicit val format: Format[SelectedPlan] = implicitly[Format[String]].inmap(SelectedPlan(_), _.value)
+
 }
