+++ cor-journey/src/main/scala/essttp/utils/EnumFormat.scala
@@ -14,14 +14,17 @@
  * limitations under the License.
  */
 
-package journey
+package essttp.utils
 
-import essttp.journey.model.JourneyId
+import enumeratum.{Enum, EnumEntry}
+import play.api.libs.json._
 
-import java.util.UUID
-import javax.inject.Singleton
-
-@Singleton
-class JourneyIdProvider {
-  def nextJourneyId(): JourneyId = JourneyId(UUID.randomUUID().toString)
+object EnumFormat {
+  def apply[T <: EnumEntry](e: Enum[T]): Format[T] = Format(
+    Reads {
+      case JsString(value) => e.withNameOption(value).map(JsSuccess(_)).getOrElse(JsError(s"Unknown ${e.getClass.getSimpleName} value: $value"))
+      case _               => JsError("Can only parse String")
+    },
+    Writes(v => JsString(v.entryName))
+  )
 }
